Small fix: lab_intro/cs225/HSLAPixel.h

+++ lab_intro/cs225/HSLAPixel.h
@@ -18,7 +18,7 @@ namespace cs225 {
     HSLAPixel (double hue, double saturation, double luminance);
     HSLAPixel (double hue, double saturation, double luminance, double alpha);
 
-    int h;  // Hue
+    double h;  // Hue
     double s;  // Saturation
     double l;  // Luminance
     double a;  // Alpha
